requirements.txt
PyYAML==6.0.1
requests==2.31.0
rdflib~=7.0.0
